Add files via upload
The Project is on the analysis of the greenhouse gases, CO2 emissions, population, and the GDP of the countries and how each country what the relationship between these factors is, and how has the relationship changed over time.
A brief introduction to the project : The relationship between consumption CO2(CO2 produced on the import of products, goods from other countries for your self-use), Production CO2(CO2 generated as a result of industrial growth, and anything within a country for its sustenance) population, GDP, income bracket of the countries may seem straight forward. Are underdeveloped and developing countries facing more challenges to improving economically because of this? The scope of the project is from 1980-2018.

The conclusions that I have arrived at are better explained in the visualizations file. What I was trying to do is to analyze the relationship between all of the different variables that are not really directly visible to us. The questions I am trying to answer are the following:

a. How does the GDP per capita affect the emission and consumption of CO2? Relationship between population vs CO2 emissions from a particular region.
b. Is there pressure on developing countries to reduce their CO2 emissions from developed countries?
c. Predicted impacts on emissions by implementing policies, measures, pledges, and targets set up for net-zero emissions.
d. How are the measures, and policies implemented across countries?
e. What should be done to optimize the net CO2 emissions for different countries?